feat: Remove splash screen and init-steps when using mobile devices.
